pyproject.toml
[tool.poetry]
name = "ghost-langchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
pinecone-client = "^3.2.2"
langchain-community = "^0.0.32"
cohere = "^5.1.8"
openai = "^1.17.1"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
langchain-cohere = "^0.1.2"
langchain-openai = "^0.1.3"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
